Convolver convolves 2d scans with a kernel
The convolver class allows to produce a convolved image from a given scan and a given kernel. Two major modes of operation are supported: a) one can generate several random kernels and display them together with the results of their action on a given image, b) given a single image and a single kernel one can modify relative weights of the entries in the kernel and see the effect.